+++ R/mcmc-distributions.R
@@ -34,9 +34,16 @@
 #'    The density estimate of each chain is plotted as a violin with
 #'    horizontal lines at notable quantiles.
 #'   }
+#'   \item{\code{mcmc_dens_chains}}{
+#'    Ridgeline kernel density plots of posterior draws with chains separated
+#'    but overlaid on a single plot. In \code{mcmc_dens_overlay} parameters
+#'    appear in separate facets; in \code{mcmc_dens_chains} they appear in the
+#'    same panel and can overlap vertically.
+#'   }
 #' }
 #'
 #' @examples
+#' set.seed(9262017)
 #' # some parameter draws to use for demonstration
 #' x <- example_mcmc_draws()
 #' dim(x)
@@ -71,6 +78,7 @@
 #' color_scheme_set("pink")
 #' mcmc_hist_by_chain(x, regex_pars = "beta")
 #' }
+#'
 #' #################
 #' ### Densities ###
 #' #################
@@ -83,6 +91,8 @@
 #' mcmc_dens_overlay(x, pars = c("sigma", "beta[2]"),
 #'                   facet_args = list(nrow = 2)) +
 #'                   facet_text(size = 14)
+#' x2 <- example_mcmc_draws(params = 6)
+#' mcmc_dens_chains(x2, pars = c("beta[1]", "beta[2]", "beta[3]"))
 #' }
 #' # separate chains as violin plots
 #' color_scheme_set("green")
@@ -171,6 +181,7 @@ mcmc_dens_overlay <- function(x,
                               regex_pars = character(),
                               transformations = list(),
                               facet_args = list(),
+                              color_chains = TRUE,
                               ...,
                               trim = FALSE) {
   check_ignored_arguments(...)
@@ -181,9 +192,82 @@ mcmc_dens_overlay <- function(x,
     transformations = transformations,
     facet_args = facet_args,
     by_chain = TRUE,
+    color_chains = color_chains,
     trim = trim,
     ...
   )
+}
+
+#' @rdname MCMC-distributions
+#' @template args-density-controls
+#' @param color_chains option for whether to separately color chains.
+#' @export
+mcmc_dens_chains <- function(x, pars = character(), regex_pars = character(),
+                             transformations = list(),
+                             color_chains = TRUE,
+                             ...,
+                             bw = NULL, adjust = NULL, kernel = NULL,
+                             n_dens = NULL) {
+  check_ignored_arguments(...)
+  data <- mcmc_dens_chains_data(x, pars = pars, regex_pars = regex_pars,
+                                transformations = transformations, bw = bw,
+                                adjust = adjust, kernel = kernel,
+                                n_dens = n_dens)
+
+  n_chains <- length(unique(data$chain))
+  if (n_chains == 1) STOP_need_multiple_chains()
+
+  # An empty data-frame to train legend colors
+  line_training <- dplyr::slice(data, 0)
+
+  if (color_chains) {
+    scale_color <- scale_color_manual(values = chain_colors(n_chains))
+  } else {
+    scale_color <- scale_color_manual(
+      values = rep(get_color("m"), n_chains),
+      guide = "none")
+  }
+
+  ggplot(data) +
+    aes_(x = ~ x, y = ~ parameter, color = ~ chain,
+         group = ~ interaction(chain, parameter)) +
+    geom_line(data = line_training) +
+    ggridges::geom_density_ridges(
+      aes_(height = ~ density),
+      stat = "identity",
+      fill = NA,
+      show.legend = FALSE) +
+    labs(color = "Chain") +
+    scale_y_discrete(limits = unique(rev(data$parameter)),
+                     expand = c(0.05, .6)) +
+    scale_color +
+    yaxis_title(FALSE) +
+    xaxis_title(FALSE) +
+    grid_lines_y(color = "gray90")  +
+    theme(axis.text.y = element_text(hjust = 1, vjust = 0))
+}
+
+#' @rdname MCMC-distributions
+#' @export
+mcmc_dens_chains_data <- function(x, pars = character(),
+                                  regex_pars = character(),
+                                  transformations = list(),
+                                  ...,
+                                  bw = NULL, adjust = NULL, kernel = NULL,
+                                  n_dens = NULL) {
+  check_ignored_arguments(...)
+
+  x %>%
+    prepare_mcmc_array(pars = pars, regex_pars = regex_pars,
+                       transformations = transformations) %>%
+    melt_mcmc() %>%
+    compute_column_density(c(.data$Parameter, .data$Chain), .data$Value,
+                           interval_width = 1,
+                           bw = bw, adjust = adjust, kernel = kernel,
+                           n_dens = n_dens) %>%
+    mutate(Chain = factor(.data$Chain)) %>%
+    rlang::set_names(tolower) %>%
+    dplyr::as_tibble()
 }
 
 #' @rdname MCMC-distributions
@@ -208,6 +292,10 @@ mcmc_violin <- function(x,
     ...
   )
 }
+
+
+
+
 
 
 
@@ -269,6 +357,7 @@ mcmc_violin <- function(x,
                        transformations = list(),
                        facet_args = list(),
                        by_chain = FALSE,
+                       color_chains = FALSE,
                        geom = c("density", "violin"),
                        probs = c(0.1, 0.5, 0.9),
                        trim = FALSE,
@@ -280,14 +369,14 @@ mcmc_violin <- function(x,
 
   geom <- match.arg(geom)
   violin <- geom == "violin"
-  geom_fun <- if (by_chain)
-    "stat_density" else paste0("geom_", geom)
+  geom_fun <- if (by_chain) "stat_density" else paste0("geom_", geom)
 
   if (by_chain || violin) {
-    if (!has_multiple_chains(x))
+    if (!has_multiple_chains(x)) {
       STOP_need_multiple_chains()
-    else
-      n_chain <- num_chains(data)
+    } else {
+      n_chains <- num_chains(data)
+    }
   }
 
   aes_mapping <- if (violin) {
@@ -315,10 +404,20 @@ mcmc_violin <- function(x,
   graph <- ggplot(data, mapping = do.call("aes_", aes_mapping)) +
     do.call(geom_fun, geom_args)
 
-  if (!violin)
+  if (!violin) {
     graph <- graph + dont_expand_x_axis()
-  if (by_chain)
-    graph <- graph + scale_color_manual(values = chain_colors(n_chain))
+  }
+
+  if (by_chain) {
+    if (color_chains) {
+      scale_color <- scale_color_manual(values = chain_colors(n_chains))
+    } else {
+      scale_color <- scale_color_manual(
+        values = rep(get_color("m"), n_chains),
+        guide = "none")
+    }
+    graph <- graph + scale_color
+  }
 
   if (n_param == 1) {
     graph <-
@@ -331,7 +430,6 @@ mcmc_violin <- function(x,
       facet_args[["scales"]] <- "free"
     graph <- graph + do.call("facet_wrap", facet_args)
   }
-
 
   graph +
     dont_expand_y_axis(c(0.005, 0)) +